+++ c10/cuda/CUDAMallocAsyncAllocator.cpp
@@ -11,6 +11,7 @@
 
 namespace c10::cuda::CUDACachingAllocator::CudaMallocAsync {
 
+using namespace c10::CachingAllocator;
 using namespace c10::CachingDeviceAllocator;
 
 #if CUDA_VERSION >= 11040
