+++ torch/csrc/xpu/Module.cpp
@@ -212,9 +212,9 @@ PyObject* THXPModule_memoryStats(PyObject* self, PyObject* arg) {
   TORCH_CHECK(THPUtils_checkLong(arg), "invalid argument to memory_stats");
   const auto device_index = THPUtils_unpackDeviceIndex(arg);
 
-  using c10::Stat;
-  using c10::StatArray;
-  using c10::StatType;
+  using c10::CachingAllocator::Stat;
+  using c10::CachingAllocator::StatArray;
+  using c10::CachingAllocator::StatType;
   using c10::CachingDeviceAllocator::DeviceStats;
 
   const auto statToDict = [](const Stat& stat) {
